modified:   src/http.c

+++ src/http.c
@@ -69,11 +69,11 @@ static void progress_fn(struct mg_connection *c, int ev, void *ev_data, void *fn
                   "POST %s HTTP/1.0\r\n"
                   "Host: %.*s:%d\r\n"
                   "Content-Type: application/octet-stream\r\n"
-                  "Content-Length: %zu\r\n"
+                  "Content-Length: %d\r\n"
                   "\r\n"
                   "\0\0\0\0",
                   mg_url_uri(req->url),
-                  (int)host.len, host.ptr, req->port, sizeof(int));
+                  (int)host.len, host.ptr, req->port, (int)sizeof(int));
 
         // append the "POST" data buffer
         // mg_send(c, &(req->progress), sizeof(int));
